+++ NEWS.md
@@ -7,6 +7,7 @@
 * Added "Stdev-fold" tab.
 * Fix: differential expression between metacells did not work with most recent version of `qs`.
 * Added Lasso selection in Annotation and Differential Expression tabs.
+* Added "# of metacells with significant inner-fold" to the "QC" tab.
 
 # MCView 0.2.9